pyproject.toml
[tool.poetry]
name = "parcelizer"
version = "0.1.0"
description = "Local-first parcel boundary extraction tool"
authors = ["Gabriel Hansen"]
readme = "README.md"
packages = [{include = "parcelizer"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
openai = "^1.50.0"
httpx = "^0.27.0"
pillow = "^10.0.0"
pytesseract = "^0.3.10"
shapely = "^2.0.0"
geopandas = "^0.14.0"
python-dotenv = "^1.0.0"
click = "^8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-vcr = "^1.0.2"
ruff = "^0.6.0"
black = "^24.0.0"

[tool.poetry.scripts]
parcelizer = "parcelizer.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "YTT", "ANN", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["ANN101", "ANN102", "S101", "PLR0913"] 